Remove cryptography from Core Applications
Closes #570

Removing, as the course is too in depth on cryptography without teaching other important areas of security.

CS2013 specifies that the undergraduate CS curriculum include the Knowledge Area Information Assurance and Security (IAS). This knowledge area includes an introduction to cryptography. But unless students take an elective course in Cryptography, they need only demonstrate a familiarity with the topic, vocabulary, the use of primes in cryptography and how public keys are used.

As pointed out by @MohamedMandouh, this course is an advanced class offered at Stanford.

Meanwhile, IAS specifies a number of other important topics, which this class does not address.